web-ui/backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
qdrant-client
psutil
sentence-transformers
beautifulsoup4